requirements.txt
constructs>=10.0.0
aws-cdk-lib>=2.0.0
aws-cdk.aws-lambda-python-alpha>=2.51.1a0